requirements.txt
pythonocc-core
numpy
matplotlib